requirements/mac.txt
accelerate==1.13.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioop-lts==0.2.2
audioread==3.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
decorator==5.2.1
fastapi==0.135.1
ffmpeg-python==0.2.0
filelock==3.25.1
fsspec==2026.2.0
future==1.0.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.6.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
lazy-loader==0.5
librosa==0.11.0
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
prometheus_client==0.24.1
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
soundfile==0.13.1
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
httptools==0.7.1
pytest==8.4.2
uvicorn==0.41.0
uvloop==0.22.1
urllib3==2.6.3
watchfiles==1.1.1
websockets==16.0
